% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_max_consensus.R
\name{cc_get_max_consensus}
\alias{cc_get_max_consensus}
\title{Get clusters which reach max consensus}
\usage{
cc_get_max_consensus(k, cluster_ward, cluster_other, return_list = FALSE)
}
\arguments{
\item{k}{[int] a vector containing the number of clusters for Ward and
for Complete-linkage (or Single-linkage) algorithms, respectively}

\item{cluster_ward}{an object of class hclust for the Ward algorithm}

\item{cluster_other}{an object of class hclust for the
Complete-linkage (or Single-linkage) algorithm}

\item{return_list}{[lgl] If TRUE the list of the elements belonging to
each cluster and the contingency table of the clustering are
shown. Otherwise, a numeric vector of lenght one reporting the
number of element classified.}
}
\value{
If return_list is FALSE (default) the number of elements
        classified.
        If return_list is TRUE, a list with the following elements:
          \item{beta_list}{
            list of the elements belonging to each cluster
          };
          \item{A_star}{contingency table of the clustering}.
}
\description{
Computes the consensus between Ward's minimum variance and
Complete-linkage (or Single-linkage) algorithms (i.e., the number of
elements classified together by both algorithms).
}
\examples{
library(CrossClustering)

data(toy)

### toy is transposed as we want to cluster samples (columns of the
### original matrix)
toy_dist <- t(toy) \%>\% dist(method = "euclidean")

### Hierarchical clustering
cluster_ward    <- toy_dist \%>\% hclust(method = "ward.D")
cluster_other <- toy_dist \%>\% hclust(method = "complete")


### cc_get_max_consensus
CrossClustering:::cc_get_max_consensus(c(3, 4),
  cluster_ward,
  cluster_other
)

}
\references{
Tellaroli P, Bazzi M., Donato M., Brazzale A. R., Draghici S. (2016).
Cross-Clustering: A Partial Clustering Algorithm with Automatic
Estimation of the Number of Clusters. PLoS ONE 11(3):   e0152333.
doi:10.1371/journal.pone.0152333
}
\author{
Paola Tellaroli, <paola [dot] tellaroli [at] unipd [dot] it>;;
Marco Bazzi, <bazzi [at] stat [dot] unipd [dot] it>;
Michele Donato, <mdonato [at] stanford [dot] edu>.
}
