% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chain_effect}
\alias{chain_effect}
\title{A toy dataset for illustrating the chain effect.}
\format{
A data frame with 28 rows and 2 variables:

\describe{
\item{\code{X}}{num} x coordinates
0 is negative.
\item{\code{Y}}{num} y coordinates.
}
}
\usage{
chain_effect
}
\description{
A toy dataset for illustrating the chain effect.
}
\keyword{datasets}
