% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossClustering-package.R
\docType{package}
\name{CrossClustering-package}
\alias{CrossClustering}
\alias{CrossClustering-package}
\title{CrossClustering: A Partial Clustering Algorithm}
\description{
Provide the 'CrossClustering' algorithm (Tellaroli et al. (2016) \doi{10.1371/journal.pone.0152333}), which is a partial clustering algorithm that combines the Ward's minimum variance and Complete Linkage algorithms, providing automatic estimation of a suitable number of clusters and identification of outlier elements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=CrossClustering}
  \item Report bugs at \url{https://github.com/CorradoLanera/CrossClustering/issues}
}

}
\author{
\strong{Maintainer}: Paola Tellaroli \email{paola.tellaroli@gmail.com}

Authors:
\itemize{
  \item Marco Bazzi \email{bazzi@stat.unipd.it}
  \item Michele Donato \email{mdonato@stanford.edu}
  \item Livio Finos \email{livio.finos@unipd.it}
  \item Philippe Courcoux \email{philippe.courcoux@oniris-nantes.fr}
  \item Corrado Lanera \email{corrado.lanera@unipd.it}
}

}
\keyword{internal}
