% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneinlista.R
\name{geneinlista}
\alias{geneinlista}
\title{geneinlista: provides - from a list of elements for each
cluster - the list of clusters to which the elements belong.}
\usage{
geneinlista(mygene, lista)
}
\arguments{
\item{mygene}{elements to be clustered}

\item{lista}{list of clustered elements}
}
\value{
The list of clusters to which the elements belong.
}
\description{
geneinlista: provides - from a list of elements for each
cluster - the list of clusters to which the elements belong.
}
\examples{
library(CrossClustering)

### Generate simulated data
toy <- matrix(NA, nrow = 10, ncol = 7)
colnames(toy) <- paste("Sample", 1:ncol(toy), sep = "")
rownames(toy) <- paste("Gene"  , 1:nrow(toy), sep = "")
set.seed(123)

toy[, 1:2] <- rnorm(n = nrow(toy) * 2, mean = 10, sd  = 0.1)
toy[, 3:4] <- rnorm(n = nrow(toy) * 2, mean = 20, sd  = 0.1)
toy[, 5:6] <- rnorm(n = nrow(toy) * 2, mean = 5 , sd  = 0.1)
toy[,   7] <- runif(n = nrow(toy)    , min  = 0 , max = 1  )

### toy is transposed as we want to cluster samples (columns of the original
### matrix)
d <- dist(t(toy), method = "euclidean")

### Run CrossClustering
toyres <- CrossClustering(d,
  k.w.min = 2,
  k.w.max = 5,
  k.c.max = 6,
  out     = TRUE
)

### geneinlista
sapply(seq_len(ncol(toy)), CrossClustering:::geneinlista, toyres$Cluster.list)

}
\references{
Tellaroli P, Bazzi M., Donato M., Brazzale A. R., Draghici S. (2016).
Cross-Clustering: A Partial Clustering Algorithm with Automatic Estimation
of the Number of Clusters. PLoS ONE 11(3):   e0152333.
doi:10.1371/journal.pone.0152333
}
\author{
Paola Tellaroli, <paola [dot] tellaroli [at] unipd [dot] it>;;
Marco Bazzi, <bazzi [at] stat [dot] unipd [dot] it>;
Michele Donato, <mdonato [at] stanford [dot] edu>.
}
