% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_zero}
\alias{is_zero}
\title{Check for zero}
\usage{
is_zero(num)
}
\arguments{
\item{num}{a numerical vector of length one}
}
\value{
a boolean, TRUE if num is 0
}
\description{
Check if a given, single, number is 0 or not
}
\examples{
CrossClustering:::is_zero(1)
CrossClustering:::is_zero(0)
}
