% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_proportion_function.R
\name{max_proportion_function}
\alias{max_proportion_function}
\title{max_proportion_function}
\usage{
max_proportion_function(k, beta.clu.ward, beta.clu.complete, dist,
  return.list = FALSE)
}
\arguments{
\item{return.list}{}
}
\value{

}
\description{
max_proportion_function
}
