% Generated by roxygen2 (4.0.2): do not edit by hand
\name{proportion.function}
\title{proportion.function}
\alias{proportion.function}
\usage{proportion.function(k, beta.clu.ward, beta.clu.complete, dist, return.list=F)}
\arguments{
\item{k}{elements}
\item{beta.clu.ward}{hclust object from Ward's method}
\item{beta.clu.complete}{hclust object from Complete-linkage method}
\item{dist}{dist object}
\item{return.list}{logical value}
}
\description{Internal function.}
