% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_table.R
\name{reverse_table}
\alias{reverse_table}
\title{Reverse the process of create a contingency table}
\usage{
reverse_table(x)
}
\arguments{
\item{x}{a contingency table}
}
\value{
a list of 2 vector corresponding to the unrolled table
}
\description{
Reverse the process of create a contingency table
}
\examples{
clust_1 <- iris[, 1:4] |>
  dist() |>
  hclust() |>
  cutree(k = 3)
clust_2 <- iris[, 1:4] |>
  dist() |>
  hclust() |>
  cutree(k = 4)
cont_table <- table(clust_1, clust_2)

reverse_table(cont_table)
}
