% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignificanceARI.R
\name{SignificanceARI}
\alias{SignificanceARI}
\title{SignificanceARI}
\usage{
SignificanceARI(ground_truth, partition)
}
\arguments{
\item{ground_truth}{The actual membership of elements in clusters}

\item{partition}{The partition coming from a clustering algorithm}
}
\value{
\item{Rand}{the Rand Index}
\item{ExpectedRand}{expected value of  Rand Index}
\item{AdjustedRand}{Adjusted Rand Index}
\item{varARI}{variance of Rand Index}
\item{NARI}{NARI}
\item{p-value}{the p-value of the test}
}
\description{
This function performs a test for testing the null hypothesis of random
agreement (i.e., adjusted Rand Index equal to 0) between two partitions.
}
\examples{
### Two moons data
data(twomoons)
d <- dist(twomoons[,1:2], method = "euclidean")
CCmoons <- CrossClusteringSingle(d,k.w.max=9,k.s.max=10)
CCmoons_clusters <- sapply(1:dim(twomoons)[1], geneinlista,
CCmoons$Cluster.list)
CCmoons_clusters[CCmoons_clusters == "integer(0)"] <- 0
CCmoons_clusters <- unlist(CCmoons_clusters) + 1

SignificanceARI(twomoons[,3], CCmoons_clusters)

}
\references{
E.M. Qannari, P. Courcoux and Faye P. (2014) Significance test of the
adjusted Rand index. Application to the free sorting task, Food Quality and
Preference, (32)93-97

L. Hubert and P. Arabie (1985) Comparing partitions, Journal of
Classification, 2, 193-218.
}
\author{
Paola Tellaroli, <paola [dot] tellaroli [at] unipd [dot] it>;
Philippe Courcoux, \email{philippe.courcoux@oniris-nantes.fr}
}
