% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twomoons}
\alias{twomoons}
\title{A famous shape data set containing two clusters with two moons shapes and
outliers}
\format{
A data frame with 52 rows and 3 variables:

\describe{
\item{\code{x}}{num} x coordinates
\item{\code{y}}{num} y coordinates.
\item{\code{clusters}}{integer} cluster membership (outliers
classified as 3rd cluster).
}
}
\usage{
twomoons
}
\description{
A famous shape data set containing two clusters with two moons shapes and
outliers
}
\keyword{datasets}
