% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_cluster.R
\name{which_cluster}
\alias{which_cluster}
\title{Provides the vector of clusters' ID to which each element belong to.}
\usage{
which_cluster(cluster_list, n_elem)
}
\arguments{
\item{cluster_list}{list of clustered elements}

\item{n_elem}{total number of elements clustered}
}
\value{
An integer vector of clusters to which the elements belong (`1`
        for the outliers, ID + 1 for the others).
}
\description{
Provides the vector of clusters' ID to which each element belong to.
}
\examples{
library(CrossClustering)

data(toy)

### toy is transposed as we want to cluster samples (columns of the
### original matrix)
toy_dist <- t(toy) \%>\% dist(method = "euclidean")

### Run CrossClustering
toyres <- cc_crossclustering(toy_dist,
  k.w.min = 2,
  k.w.max = 5,
  k2.max  = 6,
  out     = TRUE
)

### which_cluster
which_cluster(toyres$Cluster.list, toyres$n.total)

}
\references{
Tellaroli P, Bazzi M., Donato M., Brazzale A. R., Draghici S. (2016).
Cross-Clustering: A Partial Clustering Algorithm with Automatic
Estimation of the Number of Clusters. PLoS ONE 11(3):   e0152333.
doi:10.1371/journal.pone.0152333
}
\author{
Paola Tellaroli, <paola [dot] tellaroli [at] unipd [dot] it>;;
Marco Bazzi, <bazzi [at] stat [dot] unipd [dot] it>;
Michele Donato, <mdonato [at] stanford [dot] edu>.
}
