% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nb_data}
\alias{nb_data}
\title{RNA-Seq dataset example}
\format{
A data frame with 100 observations on 36 numeric variables.
}
\source{
Data included in the bioconductor package \code{maSigPro}.
\url{https://doi.org/doi:10.18129/B9.bioc.maSigPro}
}
\usage{
nb_data
}
\description{
\code{nb_data} contains a subset of a bigger normalized negative binomial
simulated dataset.
}
\details{
This dataset is part of a larger simulated and normalized dataset with 2
experimental groups, 6 time-points and 3 replicates. Simulation has been done
by using a negative binomial distribution. The first 20 genes are simulated
with changes among time.
}
\keyword{datasets}
