% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermSignificanceARI.R
\name{PermSignificanceARI}
\alias{PermSignificanceARI}
\title{A permutation test for testing the null hypothesis of
random agreement (i.e., adjusted Rand Index equal to 0) between
two partitions.}
\usage{
PermSignificanceARI(ground_truth, partition)
}
\arguments{
\item{ground_truth}{[int] A vector of the actual membership of elements
in clusters}

\item{partition}{The partition coming from a clustering algorithm}
}
\value{
A data.frame with two columns:
  \item{Stat}{the adjusted Rand Index}
  \item{p-value}{the p-value of the test}
}
\description{
A permutation test for testing the null hypothesis of
random agreement (i.e., adjusted Rand Index equal to 0) between
two partitions.
}
\examples{
library(CrossClustering)

clusters <- iris[-5] \%>\%
  dist \%>\%
  hclust(method = 'ward.D') \%>\%
  cutree(k = 3)

ground_truth <- iris[[5]] \%>\% as.numeric()

PermSignificanceARI(ground_truth, clusters)

}
\references{
Samuh M. H., Leisch F., and Finos L. (2014), Tests for Random Agreement
in Cluster Analysis, Statistica Applicata-Italian Journal of Applied
Statistics, vol. 26, no. 3, pp. 219-234.

L. Hubert and P. Arabie (1985) Comparing partitions, Journal of
Classification, 2, 193-218.
}
\author{
Paola Tellaroli, <paola [dot] tellaroli [at] unipd [dot] it>;
Livio Finos, <livio [dot] finos [at] unipd [dot] it>
}
