% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worms}
\alias{worms}
\title{TITLE}
\format{A data frame with 423 rows and 6 variables:

\describe{
   \item{\code{final}}{integer} gold standard classification: 1 is positive
     0 is negative.
   \item{\code{id}}{integer} ID code as retrieved from pubmed.
   \item{\code{title}}{character} Title of the record retrieved from pubmed.
   \item{\code{year}}{integer} Year of the record retrieved from pubmed.
   \item{\code{authors}}{character} Authors of the record retrieved from
     pubmed.
   \item{\code{Abstract}}{character} Abstract of the record retrieved from
     pubmed.
}}
\usage{
worms
}
\description{
DESCRIPTION
}
\keyword{datasets}
